Delete hard references
The `deleteAllNodes()` method does not work as advertised if there
are hard references anywhere inside of the repository. If someone
is calling `deleteAllNodes()` it is likely that they are desiring
that all of the nodes be deleted, references and all.

For example, `phpcr phpcr:purge` help states:

> The phpcr:purge command remove all the non-standard nodes from
> the content repository

This was not happening if references existed in the repository.
(Possible this was depending on implementation)