Dot not escaped in path eval in Sql2Generator.php
I found this broken for my node path which was something like
/mypath/pages/23948327.3489

Im sure this fixes my problem. But its probably better to check for non-alpha-numeric characters using a regex. If so just escape it, i guess it cant hurt that much. But I leave that for the SQL2 gurus since this is my first real encounter with SQL2